Struct async_executors::exec::AsyncStd

pub struct AsyncStd;

Available on crate feature async_std only.

An executor that spawns tasks on async-std. In contrast to the other executors, this one is not self contained, because async-std does not provide an API that allows that, so the threadpool is global.

It works on Wasm.

Implementations

impl AsyncStd

pub fn new() -> Self

Create a new AsyncStd wrapper, forwards to Default::default.

pub fn block_on<F: Future>(future: F) -> F::Output

Available on non-target_os="unknown" only.

Wrapper around async_std::task::block_on. This is not available on Wasm as Wasm does not have threads and you’re not allowed to block the only thread you have.